test/lib/cql_test_env: Disable security for tests
Use the new trivial role manager and switch to "allow all" authorizer
and authenticator to disable security for tests by default. We need this
to better isolate the unit tests from other database activities.

For example, we have a test case in enterprise tree, which checks for
cache accesses during a test. As we found out, in some cases, the
permission cache might refresh itself, causing background reads and
failing the test.